refactoring invokedoperation model out to its own file for readability
