+++ internal/model/base.go
@@ -1,0 +1,24 @@
+package model
+
+import (
+	"time"
+
+	"github.com/google/uuid"
+	"gorm.io/gorm"
+)
+
+// Base contains common columns for all tables.
+type BaseWithGuidPrimaryKey struct {
+	ID        uuid.UUID `gorm:"type:uuid;primary_key;"`
+	CreatedAt time.Time
+	UpdatedAt time.Time
+	DeletedAt *time.Time `sql:"index"`
+}
+
+// BeforeCreate will set a UUID rather than numeric ID.
+func (base *BaseWithGuidPrimaryKey) BeforeCreate(tx *gorm.DB) error {
+	if base.ID == uuid.Nil {
+		base.ID = uuid.New()
+	}
+	return nil
+}
